Added Lifts & Changing Nodes
Lifts where added to the game. Lifts float to the sides of the level, allowing you to fly back to the top without losing a life.
Upon jumping on top of a new node, the node will now "level up". the goal is to max out every node. (The game doesn't detect a completed board yet)